[85980244] Add SVG icons definitions and icon partial
To use SVGs inline from a compressed file, a "definition" needs
to be included in the HTML. The "definition" is the compressed
SVG file. This needs to be an ERB file because the compiled
SVG is not valid HAML.

The _icon.html.haml partial is needed to render an individual
icon inline.