[90552958] Any node can have or not have a detail chart
We initially had a rule that a detail chart was displayed based on the node
type. E.g. measure_source and category nodes did not have a detail chart,
but every other type did. More recently, we've decided that any node type
may have a detail chart, if it's specified in the PUBLIC_CHARTS_TREE.